refactor(cli): rename annotation bulk-delete to entity-first <entity>-annotations delete
Restructure the three unshipped annotation bulk-delete commands from
`px <entity> delete-annotations` to entity-first form
`px <entity>-annotations delete`, introducing `span-annotations`,
`trace-annotations`, and `session-annotations` as top-level command
groups (matching the existing `annotation-config` precedent).

Each new entity lives in its own factory file under `src/commands/`;
the pre-existing `spanAnnotations.ts` / `traceAnnotations.ts`
fetch-helper modules are unchanged. Flag set, authorization gate
(`--all` XOR `[--start-time, --end-time)`), payload shape, exit codes,
and pre-flight validation order are byte-identical to the removed
leaves. No backwards-compat aliases — the original commands have not
shipped.

Tests, CLI README, and the phoenix-cli skill docs (SKILL.md +
axial/open-coding references) are updated to the new surface; adds a
minor changeset for @arizeai/phoenix-cli.
