+++ js/packages/phoenix-cli/src/commands/formatAnnotationDelete.ts
@@ -23,12 +23,12 @@ export interface FormatAnnotationDeleteOutputOptions {
 }
 
 /**
- * Format the structured success record returned by `delete-annotations`.
+ * Format the structured success record returned by `<entity>-annotations delete`.
  *
- * Every delete-annotations call returns a structured record that echoes the
- * authorized filter so downstream callers and audit trails can distinguish
- * `--all` from a bounded time-window delete without re-deriving it from CLI
- * flags.
+ * Every `<entity>-annotations delete` call returns a structured record that
+ * echoes the authorized filter so downstream callers and audit trails can
+ * distinguish `--all` from a bounded time-window delete without re-deriving it
+ * from CLI flags.
  */
 export function formatAnnotationDeleteOutput({
   result,
