+++ js/packages/phoenix-cli/src/commands/index.ts
@@ -7,6 +7,9 @@ export * from "./span";
 export * from "./dataset";
 export * from "./experiment";
 export * from "./session";
+export * from "./spanAnnotationsCommand";
+export * from "./traceAnnotationsCommand";
+export * from "./sessionAnnotationsCommand";
 export * from "./prompt";
 export * from "./api";
 export * from "./docs";
