+++ js/packages/phoenix-cli/test/cli.test.ts
@@ -54,13 +54,10 @@ describe("Phoenix CLI", () => {
 
     expect(traceCommand).toBeDefined();
     expect(traceCommand?.commands.map((command) => command.name())).toEqual(
-      expect.arrayContaining([
-        "list",
-        "get",
-        "annotate",
-        "add-note",
-        "delete-annotations",
-      ])
+      expect.arrayContaining(["list", "get", "annotate", "add-note"])
+    );
+    expect(traceCommand?.commands.map((command) => command.name())).not.toContain(
+      "delete-annotations"
     );
     expect(
       program.commands.find((command) => command.name() === "traces")
@@ -75,12 +72,10 @@ describe("Phoenix CLI", () => {
 
     expect(spanCommand).toBeDefined();
     expect(spanCommand?.commands.map((command) => command.name())).toEqual(
-      expect.arrayContaining([
-        "list",
-        "annotate",
-        "add-note",
-        "delete-annotations",
-      ])
+      expect.arrayContaining(["list", "annotate", "add-note"])
+    );
+    expect(spanCommand?.commands.map((command) => command.name())).not.toContain(
+      "delete-annotations"
     );
     expect(
       program.commands.find((command) => command.name() === "spans")
@@ -110,14 +105,11 @@ describe("Phoenix CLI", () => {
 
     expect(sessionCommand).toBeDefined();
     expect(sessionCommand?.commands.map((command) => command.name())).toEqual(
-      expect.arrayContaining([
-        "list",
-        "get",
-        "annotate",
-        "add-note",
-        "delete-annotations",
-      ])
+      expect.arrayContaining(["list", "get", "annotate", "add-note"])
     );
+    expect(
+      sessionCommand?.commands.map((command) => command.name())
+    ).not.toContain("delete-annotations");
     const listCommand = sessionCommand?.commands.find(
       (command) => command.name() === "list"
     );
@@ -290,5 +282,41 @@ describe("Phoenix CLI", () => {
   it("should include auth in the top-level help output", () => {
     const program = createProgram();
     expect(program.helpInformation()).toContain("auth");
+  });
+
+  it("should register span-annotations as a top-level command with a delete subcommand", () => {
+    const program = createProgram();
+    const spanAnnotationsCommand = program.commands.find(
+      (command) => command.name() === "span-annotations"
+    );
+
+    expect(spanAnnotationsCommand).toBeDefined();
+    expect(
+      spanAnnotationsCommand?.commands.map((command) => command.name())
+    ).toEqual(expect.arrayContaining(["delete"]));
+  });
+
+  it("should register trace-annotations as a top-level command with a delete subcommand", () => {
+    const program = createProgram();
+    const traceAnnotationsCommand = program.commands.find(
+      (command) => command.name() === "trace-annotations"
+    );
+
+    expect(traceAnnotationsCommand).toBeDefined();
+    expect(
+      traceAnnotationsCommand?.commands.map((command) => command.name())
+    ).toEqual(expect.arrayContaining(["delete"]));
+  });
+
+  it("should register session-annotations as a top-level command with a delete subcommand", () => {
+    const program = createProgram();
+    const sessionAnnotationsCommand = program.commands.find(
+      (command) => command.name() === "session-annotations"
+    );
+
+    expect(sessionAnnotationsCommand).toBeDefined();
+    expect(
+      sessionAnnotationsCommand?.commands.map((command) => command.name())
+    ).toEqual(expect.arrayContaining(["delete"]));
   });
 });
